Rename `EventId` value field. Move ID validation into corresponding utility classes.
