+++ client/src/test/java/org/spine3/base/EventsShould.java
@@ -37,6 +37,7 @@ import static com.google.common.collect.Lists.newArrayList;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
+import static org.spine3.base.Events.checkValid;
 import static org.spine3.base.Events.getActor;
 import static org.spine3.base.Events.getMessage;
 import static org.spine3.base.Events.getProducer;
@@ -227,5 +228,15 @@ public class EventsShould {
         final EventId convertedBack = Stringifiers.fromString(str, EventId.class);
 
         assertEquals(id, convertedBack);
+    }
+
+    @Test(expected = IllegalArgumentException.class)
+    public void reject_empty_event_id() {
+        checkValid(EventId.getDefaultInstance());
+    }
+
+    @Test
+    public void accept_generated_event_id() {
+        checkValid(EventFactory.generateId());
     }
 }
